requirements.txt
streamlit
numpy
torch
torchvision
Pillow
scikit-image
opencv-python
matplotlib
pandas